% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringshaper-package.R
\docType{package}
\name{stringshaper-package}
\alias{stringshaper}
\alias{stringshaper-package}
\title{stringshaper: Bindings to the HarfBuzz library for Text Shaping}
\description{
Provides access to the text shaping functionality in the HarfBuzz
    library. 'stringshaper' is a low-level utility package mainly for graphic 
    devices that expands upon the font toolset provided by the 'systemfonts'
    package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/stringshaper}
  \item Report bugs at \url{https://github.com/r-lib/stringshaper/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomas.pedersen@rstudio.com} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
