% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textshaping-package.R
\docType{package}
\name{textshaping-package}
\alias{textshaping}
\alias{textshaping-package}
\title{textshaping: Bindings to the HarfBuzz and Fribidi libraries for Text Shaping}
\description{
Provides access to the text shaping functionality in the HarfBuzz
    library and the bidirectional algorithm in the Fribidi libary. 'textshaping' 
    is a low-level utility package mainly for graphic devices that expands upon 
    the font toolset provided by the 'systemfonts' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/textshaping}
  \item Report bugs at \url{https://github.com/r-lib/textshaping/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomas.pedersen@rstudio.com} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
